Setting firm release versions
As PHPUnit is a breaking upgrade with void return types, this commit rolls back version requirements to strict 5.5 or 5.6 candidates of Testbench and illuminate/support, as well as PHPUnit. This will go into a fix release, whilst the updated versions for Laravel 5.7 and beyond will go into a breaking change.